Refactor signal chain to work with small analysis buffer window
So far, the analysis buffer contained the complete sample data of the current
file. With live input, that won't be possible anymore. We only have access to
the current data slice, ideally sized just like our analysis window size.

This change prepares the existing code for addition of the PulseAudio input
as an alternative to SDL.