Refactor effects into modules implementing LightingEffect
This makes the Photonizer much smaller and aggregates less
hard-to-name-uniquely functions in the file. Each effect is a neat
little module now, easy to overview.

Some effects have a very visible alpha baking pass now that would be
nice to get rid of. Not sure if it can be, say, optimized out.
(The pass was already needed before, but it was less obvious.)